+++ src/modules/type-room/type-room.service.ts
@@ -5,58 +5,70 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { TypeRoom } from 'src/model';
 import { Repository } from 'typeorm';
 import { TypeRoomResponse } from './entities/type-room.entity';
+import { ApiResponse } from 'src/common/entities/typeResponse';
+import { ErrorResponseWithStatusCode } from 'src/common/entities/errorEntity';
 
 @Injectable()
 export class TypeRoomService {
   constructor(
     @InjectRepository(TypeRoom)
     private typeRoomRepository: Repository<TypeRoom>,
-  ) {}
-  async create(createTypeRoomDto: CreateTypeRoomDto): Promise<TypeRoomResponse | null> {
+  ) { }
+  async create(createTypeRoomDto: CreateTypeRoomDto): Promise<ApiResponse<null>> {
     try {
-      const result = await this.typeRoomRepository.save(createTypeRoomDto);
-      return TypeRoomResponse.fromEntity(result);
-    } catch (error) {
+      const result = await this.typeRoomRepository.create(createTypeRoomDto);
+      await this.typeRoomRepository.save(result);
+      return new ApiResponse(true, null, 'Type room created successfully', 201);
+    } catch (error: ErrorResponseWithStatusCode) {
       throw new InternalServerErrorException(error.message);
     }
   }
 
-  async findAll(): Promise<TypeRoomResponse[] | null> {
+  async findAll(): Promise<TypeRoomResponse[]> {
     try {
       const result = await this.typeRoomRepository.find();
       return result.map((item) => TypeRoomResponse.fromEntity(item));
-    } catch (error) {
+    } catch (error: ErrorResponseWithStatusCode) {
       throw new InternalServerErrorException(error.message);
     }
   }
 
-  async findOne(id: string): Promise<TypeRoomResponse | null> {
+  async findOne(id: string): Promise<TypeRoomResponse> {
     try {
       const result = await this.typeRoomRepository.findOne({ where: { id: id } });
-      return result ? TypeRoomResponse.fromEntity(result) : null;
-    } catch (error) {
+      if (!result) {
+        throw new BadRequestException(`Loại phòng (ID: ${id}) không tồn tại trong hệ thống!`);
+      }
+      return TypeRoomResponse.fromEntity(result);
+    } catch (error: ErrorResponseWithStatusCode) {
       throw new InternalServerErrorException(error.message);
     }
   }
 
-  async update(id: string, updateTypeRoomDto: UpdateTypeRoomDto): Promise<TypeRoomResponse | null> {
+  async update(id: string, updateTypeRoomDto: UpdateTypeRoomDto): Promise<ApiResponse<null>> {
     try {
       const typeRoom = await this.typeRoomRepository.findOne({ where: { id: id } });
       if (!typeRoom) {
         throw new BadRequestException(`Loại phòng (ID: ${id}) không tồn tại trong hệ thống!`);
       }
       typeRoom.name = updateTypeRoomDto.name || typeRoom.name;
-      return this.typeRoomRepository.save(typeRoom);
-    } catch (error) {
+      await this.typeRoomRepository.save(typeRoom);
+      return new ApiResponse(true, null, 'Type room updated successfully', 200);
+    } catch (error: ErrorResponseWithStatusCode) {
       throw new InternalServerErrorException(error.message);
     }
   }
 
   async remove(id: string) {
     try {
-      const result = await this.typeRoomRepository.delete(id);
-      return result;
-    } catch (error) {
+      const typeRoom = await this.typeRoomRepository.findOne({ where: { id: id } });
+      if (!typeRoom) {
+        throw new BadRequestException(`Loại phòng (ID: ${id}) không tồn tại trong hệ thống!`);
+      }
+
+      await this.typeRoomRepository.delete(id);
+      return new ApiResponse(true, null, 'Type room deleted successfully', 200);
+    } catch (error: ErrorResponseWithStatusCode) {
       throw new InternalServerErrorException(error.message);
     }
   }
